app_tkinter/requirements.txt
# Dependencias para la aplicación móvil Tkinter
# Compatible con Python 3.8+ incluyendo Python 3.13

# Framework GUI mejorado
customtkinter==5.2.1

# Manejo de JWT
PyJWT==2.9.0

# Requests HTTP
requests==2.32.3

# Parseo y manejo de XML
lxml==5.3.0

# Visualización de gráficas
matplotlib==3.9.2
plotly==5.24.1

# Manejo de imágenes
Pillow>=11.0.0

# Utilidades
python-dateutil==2.9.0
pytz==2024.1